requirements.txt

# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

ultralytics>=8.0.0

opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
albumentations>=1.3.1

numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tensorboard>=2.13.0

PyQt5>=5.15.9
PyQt5-sip>=12.12.0
pyqtgraph>=0.13.3

pyyaml>=6.0
omegaconf>=2.3.0

h5py>=3.9.0
tqdm>=4.65.0
imgaug>=0.4.0

pycocotools>=2.0.6
thop>=0.1.1

loguru>=0.7.0
colorama>=0.4.6
tabulate>=0.9.0
rich>=13.4.0

higher>=0.2.1
learn2learn>=0.2.0

onnx>=1.14.0
onnxruntime>=1.15.0

jupyter>=1.0.0
ipywidgets>=8.0.0
pytest>=7.4.0

requests>=2.31.0
psutil>=5.9.0
gpustat>=1.1.1
